Refactor EventHandler import: move from model_system_event_listener to event_handler
